Moves all db-related code to db module from util
includes pep8 compliance for both modules.
